// client/src/app/models/enums/nechto-card-action.enum.ts
export enum NechtoCardAction {
    Event_Action_Flamethrower = 'flamethrower',
    Event_Action_Persistence = 'persistence',
    Event_Action_Whiskey = 'whiskey',
    Event_Action_SwapPlaces = 'swap-places',
    Event_Action_RunAway = 'run-away',
    Event_Action_Suspicion = 'suspicion',
    Event_Action_Analysis = 'analysis',
    Event_Action_Temptation = 'temptation',
    Event_Action_LookAround = 'look-around',
    Event_Action_Axe = 'axe',
    Event_Infection_It = 'it',
    Event_Infection_Infection1 = 'infection1',
    Event_Infection_Infection2 = 'infection2',
    Event_Infection_Infection3 = 'infection3',
    Event_Infection_Infection4 = 'infection4',
    Event_Defence_Fear = 'fear',
    Event_Defence_GoodHere = 'good-here',
    Event_Defence_NoBarbecue = 'no-barbecue',
    Event_Defence_Miss = 'miss',
    Event_Defence_NoThanks = 'no-thanks',
    Event_Obstacle_Quarantine = 'quarantine',
    Event_Obstacle_LockedDoor = 'locked-door',
    Panic_Friends = 'friends',
    Panic_OldRopes = 'old-ropes',
    Panic_BetweenUs = 'between-us',
    Panic_IsItParty = 'is-it-party',
    Panic_OneTwo = 'one-two',
    Panic_ThreeFour = 'three-four',
    Panic_BlindDate = 'blind-date',
    Panic_ChainReaction = 'chain-reaction',
    Panic_Forgetfulness = 'forgetfulness',
    Panic_GoAway = 'go-away',
    Panic_Oops = 'oops',
    Panic_ConfessionTime = 'confession-time'
}